shaved off ~0.2 secons

+++ src/mapped_memory.cpp
@@ -1,11 +1,7 @@
 #include <cstdint>
-#include <fstream>
 #include <iomanip>
 #include <iostream>
 #include <map>
-#include <string>
-#include <thread>
-#include <vector>
 #include <string_view>
 
 #include <fcntl.h>
@@ -24,18 +20,16 @@ struct Stats {
   int64_t cnt = 0;
 };
 
-
 using namespace std::chrono;
 
 int main() {
 
-  //unsigned int num_threads = std::thread::hardware_concurrency();
-  //if (num_threads == 0)
-    //num_threads = 4;
-
+  // unsigned int num_threads = std::thread::hardware_concurrency();
+  // if (num_threads == 0)
+  // num_threads = 4;
 
   auto file_time_s = high_resolution_clock::now();
-  
+
   int fd = open("../1brc/measurements.txt", O_RDONLY);
   struct stat sb;
   fstat(fd, &sb);
@@ -52,14 +46,12 @@ int main() {
   std::map<std::string_view, Stats> results;
   char *cursor = file_ptr;
   char *end = file_ptr + file_size;
-  
+
   auto file_time_e = high_resolution_clock::now();
   auto file_time = duration_cast<microseconds>(file_time_e - file_time_s);
 
-
   auto parse_time_s = high_resolution_clock::now();
   while (cursor < end) {
-
 
     char *semicolon = cursor;
     while (*semicolon != ';') {
@@ -70,38 +62,41 @@ int main() {
     while (*newline != '\n') {
       newline++;
     }
-    
+
     std::string_view city(cursor, semicolon - cursor);
 
     int16_t temp = 0;
     bool negative = false;
     for (char *curr = semicolon + 1; curr < newline; ++curr) {
-      if (*curr == '-') negative = true;
+      if (*curr == '-')
+        negative = true;
       else if (*curr >= '0' && *curr <= '9') {
         temp = temp * 10 + (*curr - '0');
       }
     }
-    if (negative) temp = -temp;
+    if (negative)
+      temp = -temp;
 
     auto &s = results[city];
-    if (temp < s.min) s.min = temp;
-    if (temp > s.max) s.max = temp;
+    if (temp < s.min)
+      s.min = temp;
+    if (temp > s.max)
+      s.max = temp;
     s.sum += temp;
     s.cnt++;
 
     cursor = newline + 1;
   }
-  
+
   auto parse_time_e = high_resolution_clock::now();
   auto parse_time = duration_cast<microseconds>(parse_time_e - parse_time_s);
-
 
   std::cout << std::fixed << std::setprecision(1) << "{";
   for (auto it = results.begin(); it != results.end(); ++it) {
     const auto &[name, s] = *it;
-    std::cout << name << "=" << s.min / 10.0f << "/" << (s.sum / s.cnt) / 10.0f
-              << "/" << s.max / 10.0f
-              << (std::next(it) == results.end() ? "" : ",");
+    std::cout << name << "=" << s.min / 10.0f << "/"
+              << (s.sum / static_cast<float>(s.cnt)) / 10.0f << "/"
+              << s.max / 10.0f << (std::next(it) == results.end() ? "" : ",");
   }
   std::cout << "}\n";
